refactor: split guest-download modules
